feat(comments): enforce delete policy + refine forbidden error

+++ src/feature/comments/application/use-cases/DeleteComment.ts
@@ -1,6 +1,7 @@
 import { type CommentRepository } from '../../domain/ports/CommentRepository'
 import { CommentNotFoundError } from '../../domain/errors/CommentNotFoundError'
 import { CommentForbiddenError } from '../../domain/errors/CommentForbiddenError'
+import { DELETE_BLOCKED_HAS_REPLIES } from '@shared/domain/errors/authorization-error-codes'
 
 export class DeleteComment {
   private readonly commentRepository: CommentRepository
@@ -18,6 +19,14 @@ export class DeleteComment {
 
     if (comment.authorId !== authorId) {
       throw new CommentForbiddenError()
+    }
+
+    const otherRepliesCount = await this.commentRepository.countRepliesByParentId(id, comment.authorId)
+    if (otherRepliesCount > 0) {
+      throw new CommentForbiddenError(
+        'Cannot delete comment: other users have replied',
+        DELETE_BLOCKED_HAS_REPLIES
+      )
     }
 
     await this.commentRepository.delete(id)
